% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtcars.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{Car parameters}
\format{Data frame with 32 rows and 11 variables
\describe{
\item{mpg}{Miles per galon}
\item{cyl}{Number of cylinders}
\item{disp}{Displacement (cu. in.)}
\item{hp}{Gross horsepower}
\item{drat}{Rear axle ratio}
\item{wt}{Weight (1000 lbs)}
\item{qsex}{1/4 mile time}
\item{vs}{V/S}
\item{am}{Transmission (0 = automatic, 1 = manual)}
\item{am}{Transmission (0 = automatic, 1 = manual)}
\item{gear}{Number of forward gears}
\item{carb}{Number of carburetors}
}}
\source{
system.file('data/Rdata.rds', package = 'datasets')
}
\usage{
mtcars
}
\description{
Compare parameters of multiple cars (names described in rownames)
}
\examples{
  mtcars
}
\keyword{datasets}
